fix (service): renamed methods to better declare their function

+++ src/main/java/com/example/evolutioncodetest/task/TaskService.java
@@ -1,9 +1,12 @@
 package com.example.evolutioncodetest.task;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
+import org.springframework.web.client.RestClientResponseException;
 
 import java.util.List;
+import java.util.NoSuchElementException;
 import java.util.UUID;
 
 @Service
@@ -16,31 +19,47 @@ public class TaskService {
         this.repository = repository;
     }
 
-    public List<TaskModel> findAll(){
+    public List<TaskModel> getAllTasks(){
         return repository.findAll();
     }
 
-    public TaskModel findById(UUID id){
+    public TaskModel getTaskById(UUID id){
         return repository.findById(id).orElseThrow();
     }
 
-    public List<TaskModel> findByDescription(String description){
+    public List<TaskModel> getTaskByDescription(String description){
         return repository.findByDescriptionContains(description);
     }
 
-    public List<TaskModel> findByStatus(boolean isCompleted){
+    public List<TaskModel> getTaskByStatus(boolean isCompleted){
         return repository.findAllByIsCompleted(isCompleted);
     }
 
     public TaskModel createTask(TaskModel task){
+        if(repository.findById(task.getId()).isPresent()){
+            throw new RestClientResponseException(
+                        String.format("Task with ID %b already exists", task.getId()),
+                        HttpStatus.BAD_REQUEST,
+                        HttpStatus.BAD_REQUEST.getReasonPhrase(),
+                        null,
+                        null,
+                        null
+                    );
+        }
         return repository.save(task);
     }
 
     public TaskModel updateTask(TaskModel task){
+        if(repository.findById(task.getId()).isEmpty()){
+            throw new NoSuchElementException(String.format("Task with id %b doesn't exists", task.getId()));
+        }
         return repository.save(task);
     }
 
     public void deleteTaskById(UUID id){
+        if(repository.findById(id).isEmpty()){
+            throw new NoSuchElementException(String.format("Task with id %b doesn't exists", id));
+        }
         repository.deleteById(id);
     }
 }
